+++ backend/data/src/main/kotlin/io/tolgee/component/CurrentDateProvider.kt
@@ -90,7 +90,7 @@ class CurrentDateProvider(
 
   val localDate: LocalDate
       get() {
-        return (forcedDate ?: date).toInstant().atZone(systemDefault()).toLocalDate()
+        return date.toInstant().atZone(systemDefault()).toLocalDate()
       }
 
   override fun getNow(): Optional<TemporalAccessor> {
